Port CUDA radar engine to Metal for macOS (Apple Silicon native)
Full conversion of 13 CUDA compute kernels to Metal Shading Language:
- gpu_pipeline.metal: NEXRAD Level 2 parsing, transposition (5 kernels)
- renderer.metal: single-station, forward render, mosaic composite (6 kernels)
- volume3d.metal: 3D volume build, smooth, ray-march, cross-section (4 kernels)

Host-side MetalRenderer.mm replaces all CUDA runtime calls with Metal
command buffers, compute pipeline states, and shared-mode buffers.

Display pipeline: GLFW + CAMetalLayer + ImGui Metal backend, replacing
the CUDA-OpenGL interop path. Shaders compile from source at runtime
when Xcode is not installed (falls back from precompiled .metallib).

macOS trackpad support: pinch-to-zoom (NSEventMaskMagnify), two-finger
scroll to pan, smart zoom (double-tap toggle), cursor-anchored zoom.

Mosaic mode uses multi-pass forward rendering with shared depth buffer
(Metal's 31 buffer binding limit prevents the CUDA single-pass approach).

3D volume and cross-section rendering disabled pending host-side
dispatch debugging (shaders compile and are ready).

+++ src/nexrad/level2_parser.cpp
@@ -237,7 +237,7 @@ static std::vector<uint8_t> decompressGzip(const std::vector<uint8_t>& data) {
     int ret = Z_OK;
     while (ret == Z_OK) {
         if (strm.total_out >= output.size()) {
-            const size_t newSize = std::min(output.size() * 2, 200u * 1024u * 1024u);
+            const size_t newSize = std::min(output.size() * 2, size_t(200u * 1024u * 1024u));
             if (newSize <= output.size()) {
                 inflateEnd(&strm);
                 return {};
